Some modifications and changes.
Added FathersComponentTab.java to start inheritance
Added playbooks/PlaybookTab.form and playbooks/PlaybookTab.java that will be the GUI to interact with playbooks.
Added the new tabs to the BurpExtender.java, but will be removed soon.
Added inheritance to ConfigurationTab.java.
Removed some files from ContextMenuActionListener.java.
Added the AllocatedProjectsTab.java and AllocatedProjectsTab.form to be the GUI when interacting with the allocated projects.
Added inheritance to NewVulnerabilityTab.java and some adjustments to color stuff.